Organize code into app package with Docker support
