+++ app/main.py
@@ -3,9 +3,9 @@ from flask import Flask, render_template, request
 from langchain_google_genai import GoogleGenerativeAI
 from langchain.chains import LLMChain
 from langchain.prompts import PromptTemplate
-from db_connector import DatabaseConnector
-from llm_api_connector import LLMConnector
-from graph_rag import GraphRAG
+from .db_connector import DatabaseConnector
+from .llm_api_connector import LLMConnector
+from .graph_rag import GraphRAG
 
 app = Flask(__name__, template_folder='templates')
 
@@ -45,7 +45,4 @@ def index():
   return render_template('user_interface.html')
 
 if __name__ == '__main__':
-  app.run(debug=True)
-
-# Created/Modified files during execution:
-print("No files were created or modified during the execution of main.py")
+  app.run(host='0.0.0.0', debug=True)
