+++ docker-compose.yml
@@ -1,0 +1,8 @@
+version: '3.8'
+services:
+  app:
+    build: .
+    ports:
+      - "5000:5000"
+    environment:
+      GOOGLE_AI_KEY: ${GOOGLE_AI_KEY}
